Updating canonical URL
I was reading the docs, and emailed this link to myself using the
iOS share sheet: `https://doggo.mrkaran.dev/docs/`. When I got the
email, the link had been changed to `https://doggo.karan.dev/docs`
which is the canonical link based on the `site` setting in
`docs/astro.config.mjs`.